+++ flexmeasures/api/common/utils/args_parsing.py
@@ -17,9 +17,15 @@ including error handling.
 @parser.error_handler
 def handle_error(error, req, schema, *, error_status_code, error_headers):
     """Replacing webargs's error parser, so we can throw custom Exceptions."""
-    if error.__class__ == ValidationError:
-        # re-package all marshmallow's validation errors as our own kind (see below)
-        raise FMValidationError(message=error.messages)
+    if isinstance(error, ValidationError):
+        messages = error.messages
+
+        # Flatten custom location wrapper
+        if isinstance(messages, dict) and "args_and_json" in messages:
+            messages = messages["args_and_json"]
+
+        raise FMValidationError(message=messages)
+
     raise error